letterz_server/requirements.txt
Django==1.8.3
djangorestframework==3.1.3
ipython==3.2.1
